Set conditional memory tests skip
- Move all memory test to the separate file
- Set them to skip by default (makes testing with `go test ./...`
  more convenient)
