+++ structures_test.go
@@ -1,0 +1,169 @@
+package wmi
+
+import "time"
+
+type Win32_PerfRawData_PerfProc_Process struct {
+	IDProcess         uint32
+	WorkingSetPrivate uint64
+}
+
+type Win32_Process struct {
+	CSCreationClassName        string
+	CSName                     string
+	Caption                    *string
+	CommandLine                *string
+	CreationClassName          string
+	CreationDate               *time.Time
+	Description                *string
+	ExecutablePath             *string
+	ExecutionState             *uint16
+	Handle                     string
+	HandleCount                uint32
+	InstallDate                *time.Time
+	KernelModeTime             uint64
+	MaximumWorkingSetSize      *uint32
+	MinimumWorkingSetSize      *uint32
+	Name                       string
+	OSCreationClassName        string
+	OSName                     string
+	OtherOperationCount        uint64
+	OtherTransferCount         uint64
+	PageFaults                 uint32
+	PageFileUsage              uint32
+	ParentProcessId            uint32
+	PeakPageFileUsage          uint32
+	PeakVirtualSize            uint64
+	PeakWorkingSetSize         uint32
+	Priority                   uint32
+	PrivatePageCount           uint64
+	ProcessId                  uint32
+	QuotaNonPagedPoolUsage     uint32
+	QuotaPagedPoolUsage        uint32
+	QuotaPeakNonPagedPoolUsage uint32
+	QuotaPeakPagedPoolUsage    uint32
+	ReadOperationCount         uint64
+	ReadTransferCount          uint64
+	SessionId                  uint32
+	Status                     *string
+	TerminationDate            *time.Time
+	ThreadCount                uint32
+	UserModeTime               uint64
+	VirtualSize                uint64
+	WindowsVersion             string
+	WorkingSetSize             uint64
+	WriteOperationCount        uint64
+	WriteTransferCount         uint64
+}
+
+type Win32_PerfRawData_PerfDisk_LogicalDisk struct {
+	AvgDiskBytesPerRead          uint64
+	AvgDiskBytesPerRead_Base     uint32
+	AvgDiskBytesPerTransfer      uint64
+	AvgDiskBytesPerTransfer_Base uint32
+	AvgDiskBytesPerWrite         uint64
+	AvgDiskBytesPerWrite_Base    uint32
+	AvgDiskQueueLength           uint64
+	AvgDiskReadQueueLength       uint64
+	AvgDiskSecPerRead            uint32
+	AvgDiskSecPerRead_Base       uint32
+	AvgDiskSecPerTransfer        uint32
+	AvgDiskSecPerTransfer_Base   uint32
+	AvgDiskSecPerWrite           uint32
+	AvgDiskSecPerWrite_Base      uint32
+	AvgDiskWriteQueueLength      uint64
+	Caption                      *string
+	CurrentDiskQueueLength       uint32
+	Description                  *string
+	DiskBytesPerSec              uint64
+	DiskReadBytesPerSec          uint64
+	DiskReadsPerSec              uint32
+	DiskTransfersPerSec          uint32
+	DiskWriteBytesPerSec         uint64
+	DiskWritesPerSec             uint32
+	FreeMegabytes                uint32
+	Frequency_Object             uint64
+	Frequency_PerfTime           uint64
+	Frequency_Sys100NS           uint64
+	Name                         string
+	PercentDiskReadTime          uint64
+	PercentDiskReadTime_Base     uint64
+	PercentDiskTime              uint64
+	PercentDiskTime_Base         uint64
+	PercentDiskWriteTime         uint64
+	PercentDiskWriteTime_Base    uint64
+	PercentFreeSpace             uint32
+	PercentFreeSpace_Base        uint32
+	PercentIdleTime              uint64
+	PercentIdleTime_Base         uint64
+	SplitIOPerSec                uint32
+	Timestamp_Object             uint64
+	Timestamp_PerfTime           uint64
+	Timestamp_Sys100NS           uint64
+}
+
+type Win32_OperatingSystem struct {
+	BootDevice                                string
+	BuildNumber                               string
+	BuildType                                 string
+	Caption                                   *string
+	CodeSet                                   string
+	CountryCode                               string
+	CreationClassName                         string
+	CSCreationClassName                       string
+	CSDVersion                                *string
+	CSName                                    string
+	CurrentTimeZone                           int16
+	DataExecutionPrevention_Available         bool
+	DataExecutionPrevention_32BitApplications bool
+	DataExecutionPrevention_Drivers           bool
+	DataExecutionPrevention_SupportPolicy     *uint8
+	Debug                                     bool
+	Description                               *string
+	Distributed                               bool
+	EncryptionLevel                           uint32
+	ForegroundApplicationBoost                *uint8
+	FreePhysicalMemory                        uint64
+	FreeSpaceInPagingFiles                    uint64
+	FreeVirtualMemory                         uint64
+	InstallDate                               time.Time
+	LargeSystemCache                          *uint32
+	LastBootUpTime                            time.Time
+	LocalDateTime                             time.Time
+	Locale                                    string
+	Manufacturer                              string
+	MaxNumberOfProcesses                      uint32
+	MaxProcessMemorySize                      uint64
+	MUILanguages                              *[]string
+	Name                                      string
+	NumberOfLicensedUsers                     *uint32
+	NumberOfProcesses                         uint32
+	NumberOfUsers                             uint32
+	OperatingSystemSKU                        uint32
+	Organization                              string
+	OSArchitecture                            string
+	OSLanguage                                uint32
+	OSProductSuite                            uint32
+	OSType                                    uint16
+	OtherTypeDescription                      *string
+	PAEEnabled                                *bool
+	PlusProductID                             *string
+	PlusVersionNumber                         *string
+	PortableOperatingSystem                   bool
+	Primary                                   bool
+	ProductType                               uint32
+	RegisteredUser                            string
+	SerialNumber                              string
+	ServicePackMajorVersion                   uint16
+	ServicePackMinorVersion                   uint16
+	SizeStoredInPagingFiles                   uint64
+	Status                                    string
+	SuiteMask                                 uint32
+	SystemDevice                              string
+	SystemDirectory                           string
+	SystemDrive                               string
+	TotalSwapSpaceSize                        *uint64
+	TotalVirtualMemorySize                    uint64
+	TotalVisibleMemorySize                    uint64
+	Version                                   string
+	WindowsDirectory                          string
+}
